Separated invoice history store
